First skeleton for ModuleAllocProfiler
Abstract commonalities with IntrusiveAllocProfiler to AllocProfilerData
