+++ PerfTools/AllocMonitor/test/BuildFile.xml
@@ -44,5 +44,6 @@
     <test name="testIntrusiveAllocProfilerOutput" command="testIntrusiveAllocMonitor IntrusiveAllocProfiler 2>&amp;1 | ${LOCALTOP}/src/PerfTools/AllocMonitor/test/verifyIntrusiveAllocProfiler.py">
       <flags PRE_TEST="testIntrusiveAllocMonitor"/>
     </test>
+    <test name="TestPerfToolsModuleAllocProfiler" command="runModuleAllocProfiler.sh"/>
   </ifrelease>
 </ifrelease>
