+++ PerfTools/AllocMonitor/test/moduleAllocProfiler_cfg.py
@@ -1,0 +1,57 @@
+import FWCore.ParameterSet.Config as cms
+
+import argparse
+import sys
+
+parser = argparse.ArgumentParser(prog=sys.argv[0], description='Test ModuleAllocProfiler service.')
+parser.add_argument("--output", default="", help="Output log file pattern")
+parser.add_argument("--skipEvents", action="store_true", help="test skipping events")
+parser.add_argument("--source", action="store_true", help="show source")
+parser.add_argument("--out", action="store_true", help="show OutputModule")
+parser.add_argument("--edmodule", action="store_true", help="show a specific ed module")
+parser.add_argument("--esmodule", action="store_true", help="show a specific es module")
+args = parser.parse_args()
+
+
+process = cms.Process("TEST")
+
+process.source = cms.Source("EmptySource")
+
+process.maxEvents.input = 3
+
+process.WhatsItESProducer = cms.ESProducer("WhatsItESProducer")
+
+process.DoodadESSource = cms.ESSource("DoodadESSource")
+
+process.Thing = cms.EDProducer("ThingProducer",
+    offsetDelta = cms.int32(1)
+)
+
+process.OtherThing = cms.EDProducer("OtherThingProducer")
+
+process.thingProducer = cms.EDProducer("ThingProducer",
+                                       offsetDelta = cms.int32(100),
+                                       nThings = cms.int32(50)
+)
+
+process.get = cms.EDAnalyzer("WhatsItAnalyzer")
+process.out = cms.OutputModule("AsciiOutputModule")
+
+process.ep = cms.EndPath(process.out+process.get, cms.Task(process.WhatsItESProducer, process.DoodadESSource, process.Thing, process.OtherThing, process.thingProducer))
+
+process.add_(cms.Service("ModuleAllocProfiler",
+    moduleNames = cms.untracked.vstring(),
+    filePattern = cms.untracked.string(args.output),
+))
+
+if args.skipEvents:
+    process.ModuleAllocProfiler.nEventsToSkip = cms.untracked.uint32(2)
+
+if args.source:
+    process.ModuleAllocProfiler.moduleNames.append("source")
+if args.edmodule:
+    process.ModuleAllocProfiler.moduleNames.append("thingProducer")
+if args.esmodule:
+    process.ModuleAllocProfiler.moduleNames.append("WhatsItESProducer")
+if args.out:
+    process.ModuleAllocProfiler.moduleNames.append("out")
